Update fcl and modify refresh endpoint to work with any user
