+++ .env.example
@@ -1,6 +1,3 @@
-# FCL Dev Wallet Base URL
-BASE_URL=http://localhost:8701
-
 # The FCL Dev Wallet needs to know how to talk to the emulator.
 # Because the FCL Dev Wallet is Javascript and uses FCL we talk to the emulator
 # via the emulators grpc-http proxy
@@ -17,6 +14,8 @@ FLOW_INIT_ACCOUNTS=0
 CONTRACT_FUNGIBLE_TOKEN=0xee82856bf20e2aa6
 CONTRACT_FLOW_TOKEN=0x0ae53cb6e3f42a79
 CONTRACT_FUSD=0xf8d6e0586b0a20c7
+FLOW_AVATAR_URL=https://avatars.onflow.org/avatar/
+
+# Funding amounts on the account page
 TOKEN_AMOUNT_FLOW=100.0
 TOKEN_AMOUNT_FUSD=100.0
-FLOW_AVATAR_URL=https://avatars.onflow.org/avatar/
