+++ README.md
@@ -117,12 +117,14 @@ To use the dev wallet, configure FCL to point to the address of a locally runnin
 
 ```javascript
 import * as fcl from "@onflow/fcl"
+import {send as grpcSend} from "@onflow/transport-grpc"
 
 fcl.config()
   // Point App at Emulator
   .put("accessNode.api", "http://localhost:8080") 
   // Point FCL at dev-wallet (default port)
   .put("discovery.wallet", "http://localhost:8701/fcl/authn") 
+  .put("sdk.transport", grpcSend)
 ```
 
 ### Test harness
