+++ hooks/useConnectedAppConfig.ts
@@ -1,9 +1,15 @@
-import {useEffect, useState} from "react"
 import {WalletUtils} from "@onflow/fcl"
+import {useEffect, useState} from "react"
+import {parseScopes} from "src/scopes"
 
 export type ConnectedAppConfig = {
   type: string
-  body: Record<string, unknown>
+  body: {
+    appDomainTag?: string
+    data: unknown
+    extensions: unknown[]
+    timestamp: number
+  }
   service: Record<string, unknown>
   config: {
     services: {"OpenID.scopes": string}
@@ -26,10 +32,9 @@ export default function useConnectedAppConfig() {
     WalletUtils.ready(callback)
   }, [])
 
-  const appScopes =
+  const appScopes = parseScopes(
     connectedAppConfig?.config?.services?.["OpenID.scopes"]
-      ?.trim()
-      ?.split(/\s+/) ?? []
+  )
 
   return {connectedAppConfig, appScopes}
 }
