+++ modules.d.ts
@@ -1,5 +1,6 @@
 // Untyped dependencies
 declare module "@onflow/fcl"
+declare module "@onflow/transport-grpc"
 declare module "@onflow/util-encode-key"
 declare module "@onflow/types"
 declare module "sha3"
