+++ package.json
@@ -15,7 +15,8 @@
   "dependencies": {
     "@headlessui/react": "^1.3.0",
     "@monaco-editor/react": "^4.2.1",
-    "@onflow/fcl": "^0.0.78-alpha.4",
+    "@onflow/fcl": "^0.0.79-alpha.2",
+    "@onflow/transport-grpc": "^0.0.1",
     "@onflow/types": "^0.0.4",
     "@onflow/util-encode-key": "^0.0.2",
     "@theme-ui/color": "^0.10.0",
