+++ src/fclConfig.ts
@@ -1,4 +1,5 @@
 import {config} from "@onflow/fcl"
+import {send as grpcSend} from "@onflow/transport-grpc"
 
 export default function fclConfig(
   flowAccessNode: string,
@@ -13,4 +14,5 @@ export default function fclConfig(
     .put("0xFUNGIBLETOKENADDRESS", contractFungibleToken)
     .put("0xFLOWTOKENADDRESS", contractFlowToken)
     .put("0xFUSDADDRESS", contractFUSD)
+    .put("sdk.transport", grpcSend)
 }
