+++ src/harness/config.js
@@ -1,4 +1,5 @@
 import * as fcl from "@onflow/fcl"
+import {send as grpcSend} from "@onflow/transport-grpc"
 
 const USE_LOCAL = true
 
@@ -7,6 +8,8 @@ fcl.config()
   .put("app.detail.title", "Test Harness")
   .put("app.detail.icon", "https://placekitten.com/g/200/200")
   .put("service.OpenID.scopes", "email")
+  .put("fcl.appDomainTag", "harness-app")
+  .put("sdk.transport", grpcSend)
 
 if (USE_LOCAL) {
   // prettier-ignore
