+++ src/scopes.ts
@@ -1,0 +1,2 @@
+export const parseScopes = (scopes?: string) =>
+  scopes?.trim()?.split(/\s+/) ?? []
